+++ CHANGELOG.md
@@ -10,8 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added `color-functions` to documentation.
 ### Changed
 - `pEls` no longer needs to use `laserMarker`.
+- (DEV)Removed several unused vars.
 ### Added
 - Added `advance` parameter for both animations and property animations.
+- Added badges to the `README.md`.
 
 
 ## [0.2.4] -